+++ TemplatePr2h/.classpath
@@ -21,7 +21,7 @@
 	<classpathentry exported="true" kind="con" path="org.testng.TESTNG_CONTAINER"/>
 	<classpathentry kind="lib" path="C:/Program Files/Java/jdk-17/lib/jrt-fs.jar"/>
 	<classpathentry kind="lib" path="C:/Program Files/Java/jdk-19/lib/jrt-fs.jar"/>
-	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.7">
+	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-18">
 		<attributes>
 			<attribute name="maven.pomderived" value="true"/>
 		</attributes>
