+++ TemplatePr2h/log/testlog1.log
@@ -374,3 +374,85 @@
 2023-08-10 18:32:16,685 [main] INFO Frame1 - Test Case Executed Successfully
 2023-08-10 18:32:17,042 [main] INFO Frame1 - Browser Closed
 2023-08-10 18:32:17,044 [main] INFO Frame1 - Framework exection Stopped
+2023-08-10 18:39:58,480 [main] INFO Frame1 - Logger Started
+2023-08-10 18:40:03,550 [main] INFO Frame1 - Chrome launches
+2023-08-10 18:40:06,518 [main] INFO Frame1 - url Loaded
+2023-08-10 18:40:06,741 [main] INFO Frame1 - maximized Browser window
+2023-08-10 18:40:06,817 [main] INFO Frame1 - Test Case Execution started
+2023-08-10 18:40:16,377 [main] ERROR Frame1 - Test Case Execution Failed
+2023-08-10 18:40:16,714 [main] INFO Frame1 - Screenshot Taken
+2023-08-10 18:40:17,142 [main] INFO Frame1 - Browser Closed
+2023-08-10 18:40:19,743 [main] INFO Frame1 - Chrome launches
+2023-08-10 18:40:21,678 [main] INFO Frame1 - url Loaded
+2023-08-10 18:40:21,874 [main] INFO Frame1 - maximized Browser window
+2023-08-10 18:40:21,888 [main] INFO Frame1 - Test Case Execution started
+2023-08-10 18:40:26,881 [main] INFO Frame1 - Ok Button is working
+2023-08-10 18:40:27,224 [main] INFO Frame1 - Screenshot Taken
+2023-08-10 18:40:27,232 [main] INFO Frame1 - Test Case Executed Successfully
+2023-08-10 18:40:27,653 [main] INFO Frame1 - Browser Closed
+2023-08-10 18:40:27,653 [main] INFO Frame1 - Framework exection Stopped
+2023-08-10 20:47:55,466 [main] INFO Frame1 - Logger Started
+2023-08-10 20:48:00,587 [main] INFO Frame1 - Chrome launches
+2023-08-10 20:48:02,753 [main] INFO Frame1 - url Loaded
+2023-08-10 20:48:02,929 [main] INFO Frame1 - maximized Browser window
+2023-08-10 20:48:02,980 [main] INFO Frame1 - Test Case Execution started
+2023-08-10 20:48:12,524 [main] ERROR Frame1 - Test Case Execution Failed
+2023-08-10 20:48:12,718 [main] INFO Frame1 - Screenshot Taken
+2023-08-10 20:48:13,034 [main] INFO Frame1 - Browser Closed
+2023-08-10 20:48:15,360 [main] INFO Frame1 - Chrome launches
+2023-08-10 20:48:17,610 [main] INFO Frame1 - url Loaded
+2023-08-10 20:48:17,791 [main] INFO Frame1 - maximized Browser window
+2023-08-10 20:48:17,809 [main] INFO Frame1 - Test Case Execution started
+2023-08-10 20:48:22,122 [main] INFO Frame1 - Ok Button is working
+2023-08-10 20:48:22,421 [main] INFO Frame1 - Screenshot Taken
+2023-08-10 20:48:22,423 [main] INFO Frame1 - Test Case Executed Successfully
+2023-08-10 20:48:22,726 [main] INFO Frame1 - Browser Closed
+2023-08-10 20:48:22,726 [main] INFO Frame1 - Framework exection Stopped
+2023-08-10 21:53:09,763 [main] INFO Frame1 - Logger Started
+2023-08-10 21:53:14,924 [main] INFO Frame1 - Chrome launches
+2023-08-10 21:53:20,809 [main] INFO Frame1 - url Loaded
+2023-08-10 21:53:21,024 [main] INFO Frame1 - maximized Browser window
+2023-08-10 21:53:21,117 [main] INFO Frame1 - Test Case Execution started
+2023-08-10 21:53:22,629 [main] ERROR Frame1 - Test Case Execution Failed
+2023-08-10 21:53:22,863 [main] INFO Frame1 - Screenshot Taken
+2023-08-10 21:53:23,258 [main] INFO Frame1 - Browser Closed
+2023-08-10 21:53:25,730 [main] INFO Frame1 - Chrome launches
+2023-08-10 21:53:27,849 [main] INFO Frame1 - url Loaded
+2023-08-10 21:53:28,032 [main] INFO Frame1 - maximized Browser window
+2023-08-10 21:53:28,054 [main] INFO Frame1 - Test Case Execution started
+2023-08-10 21:53:28,974 [main] ERROR Frame1 - Test Case Execution Failed
+2023-08-10 21:53:29,165 [main] INFO Frame1 - Screenshot Taken
+2023-08-10 21:53:29,521 [main] INFO Frame1 - Browser Closed
+2023-08-10 21:53:29,523 [main] INFO Frame1 - Framework exection Stopped
+2023-08-10 22:05:42,925 [main] INFO Frame1 - Logger Started
+2023-08-10 22:05:47,433 [main] INFO Frame1 - Chrome launches
+2023-08-10 22:05:49,766 [main] INFO Frame1 - url Loaded
+2023-08-10 22:05:49,969 [main] INFO Frame1 - maximized Browser window
+2023-08-10 22:05:50,044 [main] INFO Frame1 - Test Case Execution started
+2023-08-10 22:05:51,630 [main] ERROR Frame1 - Test Case Execution Failed
+2023-08-10 22:05:51,803 [main] INFO Frame1 - Screenshot Taken
+2023-08-10 22:05:52,102 [main] INFO Frame1 - Browser Closed
+2023-08-10 22:05:54,854 [main] INFO Frame1 - Chrome launches
+2023-08-10 22:05:57,084 [main] INFO Frame1 - url Loaded
+2023-08-10 22:05:57,253 [main] INFO Frame1 - maximized Browser window
+2023-08-10 22:05:57,273 [main] INFO Frame1 - Test Case Execution started
+2023-08-10 22:05:58,257 [main] ERROR Frame1 - Test Case Execution Failed
+2023-08-10 22:05:58,453 [main] INFO Frame1 - Screenshot Taken
+2023-08-10 22:05:58,704 [main] INFO Frame1 - Browser Closed
+2023-08-10 22:05:58,705 [main] INFO Frame1 - Framework exection Stopped
+2023-08-10 22:07:38,648 [main] INFO Frame1 - Logger Started
+2023-08-10 22:07:42,478 [main] INFO Frame1 - Chrome launches
+2023-08-10 22:07:44,707 [main] INFO Frame1 - url Loaded
+2023-08-10 22:07:44,927 [main] INFO Frame1 - maximized Browser window
+2023-08-10 22:07:44,987 [main] INFO Frame1 - Test Case Execution started
+2023-08-10 22:07:46,455 [main] ERROR Frame1 - Test Case Execution Failed
+2023-08-10 22:07:46,652 [main] INFO Frame1 - Screenshot Taken
+2023-08-10 22:07:46,935 [main] INFO Frame1 - Browser Closed
+2023-08-10 22:07:49,502 [main] INFO Frame1 - Chrome launches
+2023-08-10 22:07:51,452 [main] INFO Frame1 - url Loaded
+2023-08-10 22:07:51,634 [main] INFO Frame1 - maximized Browser window
+2023-08-10 22:07:51,652 [main] INFO Frame1 - Test Case Execution started
+2023-08-10 22:07:52,630 [main] ERROR Frame1 - Test Case Execution Failed
+2023-08-10 22:07:52,831 [main] INFO Frame1 - Screenshot Taken
+2023-08-10 22:07:53,081 [main] INFO Frame1 - Browser Closed
+2023-08-10 22:07:53,084 [main] INFO Frame1 - Framework exection Stopped
