+++ TemplatePr2h/pom.xml
@@ -14,8 +14,8 @@
 
   <properties>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <maven.compiler.source>1.7</maven.compiler.source>
-    <maven.compiler.target>1.7</maven.compiler.target>
+    <maven.compiler.source>1.18</maven.compiler.source>
+    <maven.compiler.target>1.18</maven.compiler.target>
   </properties>
 
   <dependencies>
