+++ TemplatePr2h/test-output/Suite/Test.xml
@@ -1,44 +1,20 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!-- Generated by org.testng.reporters.JUnitXMLReporter -->
-<testsuite ignored="0" hostname="DESKTOP-FV0KD14" failures="1" tests="2" name="Test" time="28.566" errors="0" timestamp="2023-08-10T18:32:17 IST">
-  <testcase classname="MavenTemplatePr2h.TemplatePr2h.TestCases" name="fileUpload" time="9.456">
-    <failure type="org.openqa.selenium.NoSuchElementException" message="no such element: Unable to locate element: {&amp;quot;method&amp;quot;:&amp;quot;xpath&amp;quot;,&amp;quot;selector&amp;quot;:&amp;quot;//input[@id=&amp;apos;loginemail&amp;apos;]&amp;quot;}
-  (Session info: chrome=115.0.5790.171)
-For documentation on this error, please visit: https://selenium.dev/exceptions/#no_such_element
-Build info: version: &amp;apos;4.3.0&amp;apos;, revision: &amp;apos;a4995e2c09*&amp;apos;
-System info: host: &amp;apos;DESKTOP-FV0KD14&amp;apos;, ip: &amp;apos;192.168.1.2&amp;apos;, os.name: &amp;apos;Windows 11&amp;apos;, os.arch: &amp;apos;amd64&amp;apos;, os.version: &amp;apos;10.0&amp;apos;, java.version: &amp;apos;19.0.1&amp;apos;
-Driver info: org.openqa.selenium.chrome.ChromeDriver
-Command: [27d0953bb85f01fc66ff21c7039aaf3f, findElement {using=xpath, value=//input[@id=&amp;apos;loginemail&amp;apos;]}]
-Capabilities {acceptInsecureCerts: false, browserName: chrome, browserVersion: 115.0.5790.171, chrome: {chromedriverVersion: 114.0.5735.90 (386bc09e8f4f..., userDataDir: C:\Users\nanos\AppData\Loca...}, goog:chromeOptions: {debuggerAddress: localhost:62039}, networkConnectionEnabled: false, pageLoadStrategy: normal, platformName: WINDOWS, proxy: Proxy(), se:cdp: ws://localhost:62039/devtoo..., se:cdpVersion: 115.0.5790.171, setWindowRect: true, strictFileInteractability: false, timeouts: {implicit: 0, pageLoad: 300000, script: 30000}, unhandledPromptBehavior: dismiss and notify, webauthn:extension:credBlob: true, webauthn:extension:largeBlob: true, webauthn:extension:minPinLength: true, webauthn:extension:prf: true, webauthn:virtualAuthenticators: true}
-Session ID: 27d0953bb85f01fc66ff21c7039aaf3f">
-      <![CDATA[org.openqa.selenium.NoSuchElementException: no such element: Unable to locate element: {"method":"xpath","selector":"//input[@id='loginemail']"}
-  (Session info: chrome=115.0.5790.171)
-For documentation on this error, please visit: https://selenium.dev/exceptions/#no_such_element
-Build info: version: '4.3.0', revision: 'a4995e2c09*'
-System info: host: 'DESKTOP-FV0KD14', ip: '192.168.1.2', os.name: 'Windows 11', os.arch: 'amd64', os.version: '10.0', java.version: '19.0.1'
-Driver info: org.openqa.selenium.chrome.ChromeDriver
-Command: [27d0953bb85f01fc66ff21c7039aaf3f, findElement {using=xpath, value=//input[@id='loginemail']}]
-Capabilities {acceptInsecureCerts: false, browserName: chrome, browserVersion: 115.0.5790.171, chrome: {chromedriverVersion: 114.0.5735.90 (386bc09e8f4f..., userDataDir: C:\Users\nanos\AppData\Loca...}, goog:chromeOptions: {debuggerAddress: localhost:62039}, networkConnectionEnabled: false, pageLoadStrategy: normal, platformName: WINDOWS, proxy: Proxy(), se:cdp: ws://localhost:62039/devtoo..., se:cdpVersion: 115.0.5790.171, setWindowRect: true, strictFileInteractability: false, timeouts: {implicit: 0, pageLoad: 300000, script: 30000}, unhandledPromptBehavior: dismiss and notify, webauthn:extension:credBlob: true, webauthn:extension:largeBlob: true, webauthn:extension:minPinLength: true, webauthn:extension:prf: true, webauthn:virtualAuthenticators: true}
-Session ID: 27d0953bb85f01fc66ff21c7039aaf3f
-at org.openqa.selenium.remote.codec.w3c.W3CHttpResponseCodec.createException(W3CHttpResponseCodec.java:200)
-at org.openqa.selenium.remote.codec.w3c.W3CHttpResponseCodec.decode(W3CHttpResponseCodec.java:133)
-at org.openqa.selenium.remote.codec.w3c.W3CHttpResponseCodec.decode(W3CHttpResponseCodec.java:53)
-at org.openqa.selenium.remote.HttpCommandExecutor.execute(HttpCommandExecutor.java:184)
-at org.openqa.selenium.remote.service.DriverCommandExecutor.invokeExecute(DriverCommandExecutor.java:167)
-at org.openqa.selenium.remote.service.DriverCommandExecutor.execute(DriverCommandExecutor.java:142)
-at org.openqa.selenium.remote.RemoteWebDriver.execute(RemoteWebDriver.java:569)
-at org.openqa.selenium.remote.ElementLocation$ElementFinder$2.findElement(ElementLocation.java:162)
-at org.openqa.selenium.remote.ElementLocation.findElement(ElementLocation.java:60)
-at org.openqa.selenium.remote.RemoteWebDriver.findElement(RemoteWebDriver.java:387)
-at org.openqa.selenium.remote.RemoteWebDriver.findElement(RemoteWebDriver.java:379)
-at org.openqa.selenium.support.pagefactory.DefaultElementLocator.findElement(DefaultElementLocator.java:70)
-at org.openqa.selenium.support.pagefactory.internal.LocatingElementHandler.invoke(LocatingElementHandler.java:39)
-at jdk.proxy2/jdk.proxy2.$Proxy19.sendKeys(Unknown Source)
-at pageLayer.UIElement.enterUsername(UIElement.java:46)
-at MavenTemplatePr2h.TemplatePr2h.TestCases.fileUpload(TestCases.java:34)
+<testsuite ignored="0" hostname="DESKTOP-FV0KD14" failures="2" tests="2" name="Test" time="14.566" errors="0" timestamp="2023-08-10T22:07:53 IST">
+  <testcase classname="testCases.TestCases" name="fileUpload" time="1.468">
+    <failure type="java.lang.NullPointerException" message="Cannot invoke &amp;quot;pageLayer.LoginPage.enterUsername(String)&amp;quot; because &amp;quot;this.login&amp;quot; is null">
+      <![CDATA[java.lang.NullPointerException: Cannot invoke "pageLayer.LoginPage.enterUsername(String)" because "this.login" is null
+at testCases.TestCases.fileUpload(TestCases.java:36)
 at java.base/java.util.ArrayList.forEach(ArrayList.java:1511)
-... Removed 28 stack frames]]>
+... Removed 25 stack frames]]>
     </failure>
   </testcase> <!-- fileUpload -->
-  <testcase classname="MavenTemplatePr2h.TemplatePr2h.TestCases" name="testOkButton" time="5.763"/>
+  <testcase classname="testCases.TestCases" name="loginTest" time="0.978">
+    <failure type="java.lang.NullPointerException" message="Cannot invoke &amp;quot;pageLayer.LoginPage.enterUsername(String)&amp;quot; because &amp;quot;this.login&amp;quot; is null">
+      <![CDATA[java.lang.NullPointerException: Cannot invoke "pageLayer.LoginPage.enterUsername(String)" because "this.login" is null
+at testCases.TestCases.loginTest(TestCases.java:21)
+at java.base/java.util.ArrayList.forEach(ArrayList.java:1511)
+... Removed 25 stack frames]]>
+    </failure>
+  </testcase> <!-- loginTest -->
 </testsuite> <!-- Test -->
