+++ TemplatePr2h/test-output/junitreports/TEST-testCases.TestCases.xml
@@ -1,44 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!-- Generated by org.testng.reporters.JUnitReportReporter -->
-<testsuite hostname="DESKTOP-FV0KD14" failures="0" tests="2" name="MavenTemplatePr2h.TemplatePr2h.TestCases" time="19.266" errors="2" timestamp="2023-08-10T21:53:29 IST" skipped="0">
-  <testcase classname="MavenTemplatePr2h.TemplatePr2h.TestCases" name="loginTest" time="6.107">
+<testsuite hostname="DESKTOP-FV0KD14" failures="0" tests="2" name="testCases.TestCases" time="14.005" errors="2" timestamp="2023-08-10T22:07:53 IST" skipped="0">
+  <testcase classname="testCases.TestCases" name="fileUpload" time="6.429">
     <error message="Cannot invoke &quot;pageLayer.LoginPage.enterUsername(String)&quot; because &quot;this.login&quot; is null" type="java.lang.NullPointerException">
       <![CDATA[java.lang.NullPointerException: Cannot invoke "pageLayer.LoginPage.enterUsername(String)" because "this.login" is null
-at MavenTemplatePr2h.TemplatePr2h.TestCases.loginTest(TestCases.java:21)
-at java.base/jdk.internal.reflect.DirectMethodHandleAccessor.invoke(DirectMethodHandleAccessor.java:104)
-at java.base/java.lang.reflect.Method.invoke(Method.java:578)
-at org.testng.internal.MethodInvocationHelper.invokeMethod(MethodInvocationHelper.java:133)
-at org.testng.internal.TestInvoker.invokeMethod(TestInvoker.java:598)
-at org.testng.internal.TestInvoker.invokeTestMethod(TestInvoker.java:173)
-at org.testng.internal.MethodRunner.runInSequence(MethodRunner.java:46)
-at org.testng.internal.TestInvoker$MethodInvocationAgent.invoke(TestInvoker.java:824)
-at org.testng.internal.TestInvoker.invokeTestMethods(TestInvoker.java:146)
-at org.testng.internal.TestMethodWorker.invokeTestMethods(TestMethodWorker.java:146)
-at org.testng.internal.TestMethodWorker.run(TestMethodWorker.java:128)
-at java.base/java.util.ArrayList.forEach(ArrayList.java:1511)
-at org.testng.TestRunner.privateRun(TestRunner.java:794)
-at org.testng.TestRunner.run(TestRunner.java:596)
-at org.testng.SuiteRunner.runTest(SuiteRunner.java:377)
-at org.testng.SuiteRunner.runSequentially(SuiteRunner.java:371)
-at org.testng.SuiteRunner.privateRun(SuiteRunner.java:332)
-at org.testng.SuiteRunner.run(SuiteRunner.java:276)
-at org.testng.SuiteRunnerWorker.runSuite(SuiteRunnerWorker.java:53)
-at org.testng.SuiteRunnerWorker.run(SuiteRunnerWorker.java:96)
-at org.testng.TestNG.runSuitesSequentially(TestNG.java:1212)
-at org.testng.TestNG.runSuitesLocally(TestNG.java:1134)
-at org.testng.TestNG.runSuites(TestNG.java:1063)
-at org.testng.TestNG.run(TestNG.java:1031)
-at org.testng.remote.AbstractRemoteTestNG.run(AbstractRemoteTestNG.java:115)
-at org.testng.remote.RemoteTestNG.initAndRun(RemoteTestNG.java:251)
-at org.testng.remote.RemoteTestNG.main(RemoteTestNG.java:77)
-]]>
-    </error>
-  </testcase> <!-- loginTest -->
-  <system-out/>
-  <testcase classname="MavenTemplatePr2h.TemplatePr2h.TestCases" name="fileUpload" time="13.159">
-    <error message="Cannot invoke &quot;pageLayer.LoginPage.enterUsername(String)&quot; because &quot;this.login&quot; is null" type="java.lang.NullPointerException">
-      <![CDATA[java.lang.NullPointerException: Cannot invoke "pageLayer.LoginPage.enterUsername(String)" because "this.login" is null
-at MavenTemplatePr2h.TemplatePr2h.TestCases.fileUpload(TestCases.java:36)
+at testCases.TestCases.fileUpload(TestCases.java:36)
 at java.base/jdk.internal.reflect.DirectMethodHandleAccessor.invoke(DirectMethodHandleAccessor.java:104)
 at java.base/java.lang.reflect.Method.invoke(Method.java:578)
 at org.testng.internal.MethodInvocationHelper.invokeMethod(MethodInvocationHelper.java:133)
@@ -69,4 +35,38 @@ at org.testng.remote.RemoteTestNG.main(RemoteTestNG.java:77)
     </error>
   </testcase> <!-- fileUpload -->
   <system-out/>
-</testsuite> <!-- MavenTemplatePr2h.TemplatePr2h.TestCases -->
+  <testcase classname="testCases.TestCases" name="loginTest" time="7.576">
+    <error message="Cannot invoke &quot;pageLayer.LoginPage.enterUsername(String)&quot; because &quot;this.login&quot; is null" type="java.lang.NullPointerException">
+      <![CDATA[java.lang.NullPointerException: Cannot invoke "pageLayer.LoginPage.enterUsername(String)" because "this.login" is null
+at testCases.TestCases.loginTest(TestCases.java:21)
+at java.base/jdk.internal.reflect.DirectMethodHandleAccessor.invoke(DirectMethodHandleAccessor.java:104)
+at java.base/java.lang.reflect.Method.invoke(Method.java:578)
+at org.testng.internal.MethodInvocationHelper.invokeMethod(MethodInvocationHelper.java:133)
+at org.testng.internal.TestInvoker.invokeMethod(TestInvoker.java:598)
+at org.testng.internal.TestInvoker.invokeTestMethod(TestInvoker.java:173)
+at org.testng.internal.MethodRunner.runInSequence(MethodRunner.java:46)
+at org.testng.internal.TestInvoker$MethodInvocationAgent.invoke(TestInvoker.java:824)
+at org.testng.internal.TestInvoker.invokeTestMethods(TestInvoker.java:146)
+at org.testng.internal.TestMethodWorker.invokeTestMethods(TestMethodWorker.java:146)
+at org.testng.internal.TestMethodWorker.run(TestMethodWorker.java:128)
+at java.base/java.util.ArrayList.forEach(ArrayList.java:1511)
+at org.testng.TestRunner.privateRun(TestRunner.java:794)
+at org.testng.TestRunner.run(TestRunner.java:596)
+at org.testng.SuiteRunner.runTest(SuiteRunner.java:377)
+at org.testng.SuiteRunner.runSequentially(SuiteRunner.java:371)
+at org.testng.SuiteRunner.privateRun(SuiteRunner.java:332)
+at org.testng.SuiteRunner.run(SuiteRunner.java:276)
+at org.testng.SuiteRunnerWorker.runSuite(SuiteRunnerWorker.java:53)
+at org.testng.SuiteRunnerWorker.run(SuiteRunnerWorker.java:96)
+at org.testng.TestNG.runSuitesSequentially(TestNG.java:1212)
+at org.testng.TestNG.runSuitesLocally(TestNG.java:1134)
+at org.testng.TestNG.runSuites(TestNG.java:1063)
+at org.testng.TestNG.run(TestNG.java:1031)
+at org.testng.remote.AbstractRemoteTestNG.run(AbstractRemoteTestNG.java:115)
+at org.testng.remote.RemoteTestNG.initAndRun(RemoteTestNG.java:251)
+at org.testng.remote.RemoteTestNG.main(RemoteTestNG.java:77)
+]]>
+    </error>
+  </testcase> <!-- loginTest -->
+  <system-out/>
+</testsuite> <!-- testCases.TestCases -->
