+++ TemplatePr2h/test-output/old/Suite/classes.html
@@ -4,18 +4,18 @@
 <th>Method name</th>
 <th>Groups</th>
 </tr><tr>
-<td>MavenTemplatePr2h.TemplatePr2h.TestCases</td>
+<td>testCases.TestCases</td>
 <td>&nbsp;</td><td>&nbsp;</td></tr>
 <tr>
 <td align='center' colspan='3'>@Test</td>
 </tr>
 <tr>
 <td>&nbsp;</td>
-<td>testOkButton</td>
+<td>fileUpload</td>
 <td>&nbsp;</td></tr>
 <tr>
 <td>&nbsp;</td>
-<td>fileUpload</td>
+<td>loginTest</td>
 <td>&nbsp;</td></tr>
 <tr>
 <td align='center' colspan='3'>@BeforeClass</td>
