+++ TemplatePr2h/test-output/old/Suite/methods.html
@@ -1,20 +1,20 @@
 <h2>Methods run, sorted chronologically</h2><h3>&gt;&gt; means before, &lt;&lt; means after</h3><p/><br/><em>Suite</em><p/><small><i>(Hover the method name to see the test class name)</i></small><p/>
 <table border="1">
 <tr><th>Time</th><th>Delta (ms)</th><th>Suite<br>configuration</th><th>Test<br>configuration</th><th>Class<br>configuration</th><th>Groups<br>configuration</th><th>Method<br>configuration</th><th>Test<br>method</th><th>Thread</th><th>Instances</th></tr>
-<tr bgcolor="ac9dc3">  <td>23/08/10 18:32:06</td>   <td>0</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;TestBase.setUp()[pri:0, instance:MavenTemplatePr2h.TemplatePr2h.TestCases@38c6f217]">&gt;&gt;setUp</td> 
-<td>&nbsp;</td>  <td>main@1510467688</td>   <td></td> </tr>
-<tr bgcolor="ac9dc3">  <td>23/08/10 18:31:48</td>   <td>-17756</td> <td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;TestBase.start()[pri:0, instance:MavenTemplatePr2h.TemplatePr2h.TestCases@38c6f217]">&gt;&gt;start</td> 
-<td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td>  <td>main@1510467688</td>   <td></td> </tr>
-<tr bgcolor="b7c064">  <td>23/08/10 18:31:56</td>   <td>-10039</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="TestCases.fileUpload()[pri:0, instance:MavenTemplatePr2h.TemplatePr2h.TestCases@38c6f217]">fileUpload</td> 
+<tr bgcolor="b7c064">  <td>23/08/10 22:07:44</td>   <td>0</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="TestCases.fileUpload()[pri:0, instance:testCases.TestCases@38c6f217]">fileUpload</td> 
   <td>main@1510467688</td>   <td></td> </tr>
-<tr bgcolor="ac9dc3">  <td>23/08/10 18:32:17</td>   <td>10786</td> <td>&nbsp;</td><td>&nbsp;</td><td title="&lt;&lt;TestBase.stop()[pri:0, instance:MavenTemplatePr2h.TemplatePr2h.TestCases@38c6f217]">&lt;&lt;stop</td> 
-<td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td>  <td>main@1510467688</td>   <td></td> </tr>
-<tr bgcolor="ac9dc3">  <td>23/08/10 18:32:16</td>   <td>10429</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&lt;&lt;TestBase.TakeDown()[pri:0, instance:MavenTemplatePr2h.TemplatePr2h.TestCases@38c6f217]">&lt;&lt;TakeDown</td> 
+<tr bgcolor="ac9dc3">  <td>23/08/10 22:07:46</td>   <td>1959</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;TestBase.setUp()[pri:0, instance:testCases.TestCases@38c6f217]">&gt;&gt;setUp</td> 
 <td>&nbsp;</td>  <td>main@1510467688</td>   <td></td> </tr>
-<tr bgcolor="ac9dc3">  <td>23/08/10 18:32:16</td>   <td>10429</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&lt;&lt;TestBase.TakeDown()[pri:0, instance:MavenTemplatePr2h.TemplatePr2h.TestCases@38c6f217]">&lt;&lt;TakeDown</td> 
-<td>&nbsp;</td>  <td>main@1510467688</td>   <td></td> </tr>
-<tr bgcolor="b7c064">  <td>23/08/10 18:32:10</td>   <td>4663</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="TestCases.testOkButton()[pri:0, instance:MavenTemplatePr2h.TemplatePr2h.TestCases@38c6f217]">testOkButton</td> 
+<tr bgcolor="b7c064">  <td>23/08/10 22:07:51</td>   <td>6671</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="TestCases.loginTest()[pri:1, instance:testCases.TestCases@38c6f217]">loginTest</td> 
   <td>main@1510467688</td>   <td></td> </tr>
-<tr bgcolor="ac9dc3">  <td>23/08/10 18:32:06</td>   <td>0</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;TestBase.setUp()[pri:0, instance:MavenTemplatePr2h.TemplatePr2h.TestCases@38c6f217]">&gt;&gt;setUp</td> 
+<tr bgcolor="ac9dc3">  <td>23/08/10 22:07:38</td>   <td>-6438</td> <td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;TestBase.start()[pri:0, instance:testCases.TestCases@38c6f217]">&gt;&gt;start</td> 
+<td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td>  <td>main@1510467688</td>   <td></td> </tr>
+<tr bgcolor="ac9dc3">  <td>23/08/10 22:07:53</td>   <td>8105</td> <td>&nbsp;</td><td>&nbsp;</td><td title="&lt;&lt;TestBase.stop()[pri:0, instance:testCases.TestCases@38c6f217]">&lt;&lt;stop</td> 
+<td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td>  <td>main@1510467688</td>   <td></td> </tr>
+<tr bgcolor="ac9dc3">  <td>23/08/10 22:07:52</td>   <td>7853</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&lt;&lt;TestBase.TakeDown()[pri:0, instance:testCases.TestCases@38c6f217]">&lt;&lt;TakeDown</td> 
+<td>&nbsp;</td>  <td>main@1510467688</td>   <td></td> </tr>
+<tr bgcolor="ac9dc3">  <td>23/08/10 22:07:52</td>   <td>7853</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&lt;&lt;TestBase.TakeDown()[pri:0, instance:testCases.TestCases@38c6f217]">&lt;&lt;TakeDown</td> 
+<td>&nbsp;</td>  <td>main@1510467688</td>   <td></td> </tr>
+<tr bgcolor="ac9dc3">  <td>23/08/10 22:07:46</td>   <td>1959</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;TestBase.setUp()[pri:0, instance:testCases.TestCases@38c6f217]">&gt;&gt;setUp</td> 
 <td>&nbsp;</td>  <td>main@1510467688</td>   <td></td> </tr>
 </table>
