+++ TemplatePr2h/test-output/old/Suite/testng.xml.html
@@ -1,1 +1,1 @@
-<html><head><title>testng.xml for Suite</title></head><body><tt>&lt;?xml&nbsp;version="1.0"&nbsp;encoding="UTF-8"?&gt;<br/>&lt;!DOCTYPE&nbsp;suite&nbsp;SYSTEM&nbsp;"https://testng.org/testng-1.0.dtd"&gt;<br/>&lt;suite&nbsp;name="Suite"&nbsp;guice-stage="DEVELOPMENT"&gt;<br/>&nbsp;&nbsp;&lt;listeners&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&lt;listener&nbsp;class-name="utility.ListenersEx"/&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&lt;listener&nbsp;class-name="utility.ExtentReporterNG"/&gt;<br/>&nbsp;&nbsp;&lt;/listeners&gt;<br/>&nbsp;&nbsp;&lt;test&nbsp;thread-count="5"&nbsp;name="Test"&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&lt;classes&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&lt;class&nbsp;name="testBase.TestBase"/&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&lt;class&nbsp;name="MavenTemplatePr2h.TemplatePr2h.TestCases"/&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&lt;/classes&gt;<br/>&nbsp;&nbsp;&lt;/test&gt;&nbsp;&lt;!--&nbsp;Test&nbsp;--&gt;<br/>&lt;/suite&gt;&nbsp;&lt;!--&nbsp;Suite&nbsp;--&gt;<br/></tt></body></html>
+<html><head><title>testng.xml for Suite</title></head><body><tt>&lt;?xml&nbsp;version="1.0"&nbsp;encoding="UTF-8"?&gt;<br/>&lt;!DOCTYPE&nbsp;suite&nbsp;SYSTEM&nbsp;"https://testng.org/testng-1.0.dtd"&gt;<br/>&lt;suite&nbsp;name="Suite"&nbsp;guice-stage="DEVELOPMENT"&gt;<br/>&nbsp;&nbsp;&lt;listeners&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&lt;listener&nbsp;class-name="utility.ListenersEx"/&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&lt;listener&nbsp;class-name="utility.ExtentReporterNG"/&gt;<br/>&nbsp;&nbsp;&lt;/listeners&gt;<br/>&nbsp;&nbsp;&lt;test&nbsp;thread-count="5"&nbsp;name="Test"&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&lt;classes&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&lt;class&nbsp;name="testBase.TestBase"/&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&lt;class&nbsp;name="testCases.TestCases"/&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&lt;/classes&gt;<br/>&nbsp;&nbsp;&lt;/test&gt;&nbsp;&lt;!--&nbsp;Test&nbsp;--&gt;<br/>&lt;/suite&gt;&nbsp;&lt;!--&nbsp;Suite&nbsp;--&gt;<br/></tt></body></html>
