+++ TemplatePr2h/test-output/old/Suite/toc.html
@@ -22,7 +22,7 @@
 </tr></table>
 <table width='100%' class='test-failed'>
 <tr><td>
-<table style='width: 100%'><tr><td valign='top'>Test (1/1/0)</td><td valign='top' align='right'>
+<table style='width: 100%'><tr><td valign='top'>Test (0/2/0)</td><td valign='top' align='right'>
   <a href='Test.html' target='mainFrame'>Results</a>
 </td></tr></table>
 </td></tr><p/>
