+++ TemplatePr2h/test-output/testng-failed.xml
@@ -7,15 +7,16 @@
   </listeners>
   <test thread-count="5" name="Test(failed)">
     <classes>
-      <class name="MavenTemplatePr2h.TemplatePr2h.TestCases">
+      <class name="testCases.TestCases">
         <methods>
+          <include name="fileUpload"/>
+          <include name="loginTest"/>
           <include name="setUp"/>
           <include name="start"/>
           <include name="TakeDown"/>
           <include name="stop"/>
-          <include name="fileUpload"/>
         </methods>
-      </class> <!-- MavenTemplatePr2h.TemplatePr2h.TestCases -->
+      </class> <!-- testCases.TestCases -->
     </classes>
   </test> <!-- Test(failed) -->
 </suite> <!-- Failed suite [Suite] -->
